Activate ignore-headlines from ox-extras
Allows for exporting just the content of an org heading (without the
text of the heading itself). This is pretty much only useful to me for
latex bibliographies, which I do entirely in latex, but would have to
put under some unrelated heading in order to avoid the double
heading from both org and latex.