Fix hackish macroexpand of diego|sp-wrap-with.
For whatever reason, I was getting the "Forgot to expand macro" error when
compiling this file. I had fixed this by using the macroexpand function (or so
I thought?) but the apparent fix either never worked or stopped working.

Defining the macro within a eval-and-compile statement seemed to do the trick.